Backend/requirements.txt
fastapi
uvicorn
authlib
sqlalchemy
asyncpg
psycopg2-binary
pyjwt
python-dotenv
gunicorn
alembic
langchain
openai